extract class DiscountData
